Moved all functions of instrument.py to noise.py to simplify
